Move selenium and sql packages into main
