+++ etiqet-selenium/src/main/java/com/neueda/etiqet/selenium/LocatorException.java
