+++ etiqet-selenium/src/main/java/com/neueda/etiqet/selenium/fixture/LocatorType.java
