+++ etiqet-sql/src/main/java/com/neueda/etiqet/sql/config/Settings.java
